Starting the split of streamlit_ui.py, better error handling
