+++ streamlit_pages/__init__.py
@@ -1,0 +1,1 @@
+# This file makes the streamlit_ui directory a Python package
